working-with-certificates
Create a Self-Signed Certificate for Code Signing

In PoweShell 3.0, the New-SelfSifgnedCertificate cmdlet was generating only SSL certificates that couldn’t be used to sign the driver and application code (unlike certificates generated by the MakeCert utility). In PowerShell 5 the new version of the New-SelfSifgnedCertificate
cmdlet can now be used to issue a Code Signing certificates.